+++ CHANGELOG.md
@@ -46,3 +46,4 @@
 ### Other changes
 
 * `Graphics2D::draw_image()` is now able to take a tuple as a position
+* When creating an image from raw bytes, the number of bytes is checked